pyproject.toml
[tool.poetry]
name = "dot-bot"
version = "0.1.0"
description = ""
authors = ["Ivan Krakovskii <iv.krak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
aiogram = "^3.3.0"
sqlalchemy = "^2.0.26"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
